refactor/#176: setting 관련 코드 feature:setting 모듈에 이전
