+++ app/build.gradle.kts
@@ -56,6 +56,7 @@ dependencies {
     implementation(projects.feature.signup)
     implementation(projects.feature.login)
     implementation(projects.feature.history)
+    implementation(projects.feature.setting)
     implementation(projects.core.designsystem)
     implementation(projects.core.ui)
     implementation(projects.core.data)
