+++ feature/home/src/main/java/com/yapp/feature/home/navigation/HomeNavigation.kt
@@ -5,21 +5,13 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.yapp.feature.home.HomeRoute
-import com.yapp.feature.home.setting.SettingRoute
 import kotlinx.serialization.Serializable
 
 @Serializable
 data object HomeRoute
 
-@Serializable
-data object SettingRoute
-
 fun NavController.navigateToHome(navOptions: NavOptions? = null) {
     navigate(HomeRoute, navOptions)
-}
-
-fun NavController.navigateToSetting(navOptions: NavOptions? = null) {
-    navigate(SettingRoute, navOptions)
 }
 
 fun NavGraphBuilder.homeNavGraph(
@@ -36,16 +28,4 @@ fun NavGraphBuilder.homeNavGraph(
     }
 }
 
-fun NavGraphBuilder.settingNavGraph(
-    navigateLogin: () -> Unit,
-    navigateBack: () -> Unit,
-    handleException: (Throwable) -> Unit,
-) {
-    composable<SettingRoute> {
-        SettingRoute(
-            navigateBack = navigateBack,
-            navigateLogin = navigateLogin,
-            handleException = handleException,
-        )
-    }
-}
+
