+++ feature/setting/build.gradle.kts
@@ -1,0 +1,16 @@
+import com.yapp.app.setNamespace
+
+plugins {
+    id("yapp.android.feature")
+}
+
+android {
+    setNamespace("feature.setting")
+}
+
+dependencies {
+    implementation(libs.androidx.core.ktx)
+    implementation(libs.androidx.appcompat)
+    implementation(libs.androidx.activity.compose)
+    implementation(libs.material)
+}
