+++ feature/setting/src/main/java/com/yapp/feature/setting/SettingContract.kt
@@ -1,4 +1,4 @@
-package com.yapp.feature.home.setting
+package com.yapp.feature.setting
 
 data class SettingState(
     val isNotificationEnabled: Boolean = false,
