+++ feature/setting/src/main/java/com/yapp/feature/setting/SettingViewModel.kt
@@ -1,4 +1,4 @@
-package com.yapp.feature.home.setting
+package com.yapp.feature.setting
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
