+++ feature/setting/src/main/res/values/strings.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="setting_screen_title">설정</string>
+    <string name="setting_screen_back_icon_content_description">뒤로 가기</string>
+    <string name="setting_screen_item_notification">알림</string>
+    <string name="setting_screen_item_terms">이용약관</string>
+    <string name="setting_screen_item_privacy_policy">개인정보 처리방침</string>
+    <string name="setting_screen_item_app_version">앱 버전</string>
+    <string name="setting_screen_item_delete_account">회원탈퇴</string>
+    <string name="setting_screen_item_logout">로그아웃</string>
+    <string name="logout_dialog_title">로그아웃 할까요?</string>
+    <string name="logout_dialog_action_button">아니요!</string>
+    <string name="logout_dialog_recommend_action_button">로그아웃</string>
+    <string name="delete_account_dialog_title">정말 탈퇴하시겠어요?</string>
+    <string name="delete_account_dialog_body">탈퇴하시면 모든 정보가 삭제돼요.</string>
+    <string name="delete_account_dialog_action_button">아니요!</string>
+    <string name="delete_account_dialog_recommend_action_button">탈퇴하기</string>
+</resources>
